refactor: extract pure evaluation functions from canvasStorage.ts
Move PricingRowOutput, RateBuildupSnapshot, isGroupActive, computeInactiveNodeIds,
snapshotFromTemplate, snapshotToCanvasDoc, canvasDocToSnapshot, evaluateSnapshot,
and computeSnapshotUnitPrice into snapshotEvaluator.ts. canvasStorage.ts re-exports
them for backwards compat. Pure functions are now importable without pulling in React.
